fix the crash bug
The pointers tria and trib in edge were not being cleared when the
triangles were cleared. This left some of them pointing to nonexistent
triangles, causing a crash when trying to find the maximum and minimum
on the edge, which requires accessing the triangle through the tria
or trib pointer.