Added comments about bugs.
If a curve lies outside the triangle containing the midpoint of its chord,
split it regardless of the values along the curve. This does not,
however, fix the stray semicircle, since the cross segment does not
intercept the contour elevation. Set BENDLIMIT to 120° to get rid of it.

Draw triangles in cyan on contour plot. Many are missing.