Occasionally repaint the window while doing a long operation.
There was a bug before this commit where, when given a breakline file (here commited; it is not final), it would flip edges up to 99% and then stay there for a long time. I added this feature partly so that I could see which few edges it was repeatedly flipping. It didn't work; the bug hid.