Fix making TIN of flattri.asc
It was checking for failure only that a new edge has both nexta and nextb
null. This happens not to detect failure in flattri.asc, which consists
of seven points whose (x,y) lie exactly on a straight line in arithmetic
progression. Added a test of goodcenter at the end of tryStartPoint.
If it can't find a good center, it can't make a good TIN.