Change fire hydrant from substring to elevation range
This is a test of the previous commit. The fire hydrant point's elevation
in the PNEZD file is 100 ft lower than it should be. Reading the criteria file
in feet correctly excludes the fire hydrant from the topo. Reading it
in meters does not exclude it and results in a topo black hole.