Set the origHash of a boldatni file
origHash is set to hash if it is not an excerpt from a single
boldatni file, that is, if it's converted from other geoid files,
from more than one file, or from nothing. It is set to the source's
origHash if it is an excerpt of a single boldatni file. This
enables the geoid file hash, which will be stored in drawings,
to be recognized as belonging to the contiguous US, even if the
file used is just a small excerpt.