Compute transforms of various sizes
Try to find the number of good terms and the noise floor. Stop when
the first half, compared to the entire previous iteration, is not
much bigger than the last half and all terms are good or noise,
or give up after 11 iterations (transform size 2048).

ITRS has some spikes in the noise floor.