Attempt to speed up polyspiral::closest
Check the spiralarcs in jumbled order, rather than linear order.
This should speed up the search, since most of the time closesofar
is less than the distance of any point to the spiralarc, but it didn't.
transmer finds the closest point on 243-, 273-, and 343-part
polyspirals to the point projected from a sphere.