Get rid of SEXAG2P2
Angles formerly formatted in SEXAG2P2 are now output in seconds
with two decimal places. Fixed with larger units is not yet implemented.
There was no way, in the previous system, to format in DMS with more
than 3 decimal places; such precision is needed when converting
grid coordinates to latitudes and longitudes.