When reading GSF, make sure file is big enough for claimed size
Fuzzing produced a 180-byte file that was read as having 111111 rows
and 11111 columns. A 180-byte file can't possibly have that many
rows and columns. Before the bugfix, this resulted in resizing
the geolattice to over 12 gigabytes, pushing other processes into
swap space, and finally aborting trying to read 4 gigs of data
from a tiny file.