When inserting a point, duplicate an adjacent bearing
This fixes the bobbypin or bedpost curve bug, at least in bezitest.
The program was inserting 0, which sometimes caused a bearing to flip
360°, resulting in a spiralarc of the polyspiral being straight.

This may still give strange results if a closed polyspiral winds
an even number of times. The nandling of the joined ends expects the
winding number to be odd (normally 1 or -1).