Delete retraces after split, as well as after splice
After splitting, it had the first point number repeated at the end,
then it split the loop at a retrace, then it split the resulting loop
consisting of just one point twice, then it joined the two single-point
loops, and finally it deleted the resulting loop of one point twice.
Now it simply deletes the retrace, if any, and the repeated point.