Check whether the points, not the indices, are different
This fixes the bug for 8 petals and DH178. 13 petals and up still
don't work.

It is legal for two indices in the polygon to refer to the same point.
When this happened, in3 was saying the point is in the triangle,
when it's one of the corners.

The corruption of points was caused by running area3((xy)poly[a],
(xy)poly[b],(xy)poly[c]) from the gdb command line.