Make sure that the triangle is inside the polygon being triangulated
After some triangulation of the 21- or 34-petal flower, the remaining
polygon had two segments in a row retraced. The program took them
to be the triangle, even though it's outside the polygon. This triangle
crossed already drawn edges, and the adjacent part of the polygon
became a notch, which wasn't triangulated.