Update CMake minimum required version to 3.5
The previous minimum required version for CMake was set to 3.1.
However, current versions of CMake warn that compatibility with CMake
versions earlier than 3.5 will be removed in a future version of CMake.
Switching to 3.5 works without requiring any other changes.

(CMake 3.5 was released on March 23, 2016, and has been widely
adopted in the industry.)