Add virtual destructor to segment class
The segment class is a polymorphic class, as it contains virtual
functions. There's also a `delete` of segment* in segment.cpp

Deleting an object of a polymorphic class type through a base class
pointer without a virtual destructor can lead to undefined behavior.
Adding a virtual destructor to the segment class ensures that the
correct destructor is called when an object is deleted through a base
class pointer, preventing potential issues and ensuring proper
resource cleanup.