Cast pointers to void* before passing to printf
This addresses a somewhat superfluous warning which pertains to the
lack of a guarantee in the C++ standard that pointers to different
types aren't guaranteed to be the same size (though they are on all
modern architectures, except for some cases where function pointers
are different sizes from non-function pointers).

While the warning is likely not relevant on any modern system, if it
were to be disabled with `-Wno-format` it would disable other printf
warnings that actually are important.  So static casting to void* is
worth it.