Avoid comparison of (char < 128)
char is a type that is not defined as either signed or unsigned (though
often signed, on Android it is unsigned, for whatever reason).

This makes it poor for handling arbitrary bytes in files when the byte
values are being examined (e.g. UTF-8 processing).  A warning caught
that the signed character type was being compared as being less than
128, which it always is.

Here the buffer is changed to hold unsigned char, necessitating some
casts to use reading functions.